+++ CHANGES.rst
@@ -14,6 +14,7 @@ New features:
     - Show a "One level up" button in the result set in browse mode.
     - Show folder icons in in the result set front of items instead a arrow behind the items.
     - Filter out non-selectable and non-folderish items in the result set when in browse mode.
+    - Add option to scan the selected list of items for other patterns.
 
 Bug fixes: